Reorganized files, and completed graph exercises
