--- src/Assignment2/Employee.java
@@ -1,14 +1,0 @@
-package Assignment2;
-
-import java.util.List;
-
-public class Employee {
-    public String name;
-    public String title;
-    public List<Employee> directReports;
-
-    Employee(String name, String title) {
-        this.name = name;
-        this.title = title;
-    }
-}
