--- src/Assignment2/OrganizationStructure.java
@@ -1,9 +1,0 @@
-package Assignment2;
-
-public class OrganizationStructure {
-    public Employee ceo;
-
-    OrganizationStructure(Employee ceo) {
-        this.ceo = ceo;
-    }
-}
