--- src/Assignment2/Tree.java
@@ -1,9 +1,0 @@
-package Assignment2;
-
-public class Tree {
-    public TreeNode root;
-
-    public Tree(TreeNode root) {
-        this.root = root;
-    }
-}
