+++ src/Assignment2/graphs/GraphNode.java
@@ -1,0 +1,8 @@
+package Assignment2.graphs;
+
+public class GraphNode {
+    public int data;
+    GraphNode(int data) {
+        this.data = data;
+    }
+}
