+++ src/Assignment2/graphs/GraphWithAdjacencyList.java
@@ -1,0 +1,8 @@
+package Assignment2.graphs;
+
+import java.util.List;
+import java.util.Map;
+
+public class GraphWithAdjacencyList {
+    protected Map<GraphNode, List<GraphNode>> adjNodes;
+}
