+++ src/Assignment2/trees/OrganizationStructure.java
@@ -1,0 +1,9 @@
+package Assignment2.trees;
+
+public class OrganizationStructure {
+    public Employee ceo;
+
+    OrganizationStructure(Employee ceo) {
+        this.ceo = ceo;
+    }
+}
