+++ src/Assignment2/trees/Tree.java
@@ -1,0 +1,9 @@
+package Assignment2.trees;
+
+public class Tree {
+    public TreeNode root;
+
+    public Tree(TreeNode root) {
+        this.root = root;
+    }
+}
