+++ src/Assignment2/trees/TreeNode.java
@@ -1,0 +1,17 @@
+package Assignment2.trees;
+
+public class TreeNode {
+    public int data;
+    public TreeNode left;
+    public TreeNode right;
+
+    public TreeNode(int data, TreeNode left, TreeNode right) {
+        this.data = data;
+        this.left = left;
+        this.right = right;
+    }
+
+    public TreeNode(int data) {
+        this.data = data;
+    }
+}
